MIX-124: use onboarding form rather than old form
